Fix issues with polynomfit

+++ include/IO/Reader.h
@@ -1,7 +1,7 @@
 #ifndef READER_H
 #define READER_h
 
-double** ReadFile(const char* fileName, int* outRowCount, int* outColumnCount);
+long double** ReadFile(const char* fileName, int* outRowCount, int* outColumnCount);
 
 double* ReadTestLine(const char* fileName);
